bench: pass dtype to GMMXX so the spherical bench measures the dtype it claims
The cuda/triton bench's shape-grid had labels like '4194304,16,32,fp16'
but the cells were all measuring the auto-promoted fp32 cuBLAS path
because GMMXX's default _compute_dtype_for_input casts fp16/bf16 input
to fp32 unless the user passes dtype= explicitly. The triton-forced
comparison call (_bench_triton_forced) goes straight to
batch_gmm_Spherical_torch_native and preserves the input dtype, so the
old harness was effectively comparing GMMXX(auto-fp32) vs torch_native
(actual-dtype) — apples-to-oranges.

Pass dtype=dtype to both GMMXX(backend='cuda') and GMMXX(backend='triton')
so the cuda path also runs at the input dtype. The xl_grid metric drops
from ~0.88 to ~0.46 with no kernel changes — the previous values were
measuring a different code path than the labels claimed.

+++ benchmarks/benchmark_cuda_vs_triton_spherical.py
@@ -69,18 +69,22 @@ def _bench_one(backend: str, N: int, D: int, K: int, dtype: torch.dtype, n_iter:
     torch.manual_seed(0)
     device = "cuda" if backend in {"cuda", "triton"} else "cpu"
     x = torch.randn(N, D, device=device, dtype=dtype)
+    # Pass dtype= so GMMXX preserves the input dtype rather than auto-promoting
+    # fp16/bf16 to fp32 internally. Otherwise this harness's "fp16"/"bf16"
+    # labels are misleading — it would measure the fp32 cuBLAS chunked path
+    # for both the explicit-fp32 and the auto-promoted-from-fp16 cells.
     # Multiple warmup fits to amortize CUDA kernel cache, allocator state.
     for _ in range(3):
         GMMXX(
             n_components=K, max_iter=2, tol=0, random_state=0,
-            covariance_type="spherical", backend=backend,
+            covariance_type="spherical", backend=backend, dtype=dtype,
         ).fit(x)
     if device == "cuda":
         torch.cuda.synchronize()
     t0 = time.perf_counter()
     GMMXX(
         n_components=K, max_iter=n_iter, tol=0, random_state=0,
-        covariance_type="spherical", backend=backend,
+        covariance_type="spherical", backend=backend, dtype=dtype,
     ).fit(x)
     if device == "cuda":
         torch.cuda.synchronize()
